GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Dependencies (refs/gitbutler/Dependencies)
   branch head: 77b15ae2a221f57d4da83128ee8a4974af3ea3d5
   - package-lock.json
   - package-lock.json
 - Content (refs/gitbutler/Content)
   - vitepress/articles/node-native-dot-env-quirks.md
   - vitepress/notes/2023-12-15-developing-for-the-20-percent.md
   - vitepress/articles/javascript-week-number.md
   - vitepress/articles/input-event-keyboard-or-mouse.md
   - vitepress/notes/moving-away-from-sass.md
   - vitepress/articles/language-direction.md
   - vitepress/articles/idb-keyval.md
   - vitepress/notes/2024-01-05-buying-non-compliance.md
 - Open Graph (refs/gitbutler/Open-Graph)
   branch head: 6598c8241ecf56d84fa62afd6067218c7409eecc
   - vitepress/.vitepress/utils/pagedata.mjs
   - vitepress/.vitepress/config.mjs
 - Lint (refs/gitbutler/Lint)
   - vitepress/notes.data.mjs
 - Triage (refs/gitbutler/Triage)
   - vitepress/.vitepress/utils/head.mjs
   - vitepress/.vitepress/config.mjs
 - Various (refs/gitbutler/Various)
   branch head: c70fdd0793766ee605dd6fb01a48e2aaefa9177d

Your previous branch was: refs/heads/main

The sha for that commit was: 8852c9302eae863a944cf61cbccecbd5ab9eca84

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch